remove old files, prep for relaunch
- housekeeping of old files and scripts for the assumption of mysql
- removed extrudeFlatGraph and old index css
