+++ scripts/!quickStartLin/1-launchDbService.sh
@@ -1,0 +1,4 @@
+ECHO Starting MySql server
+sudo systemctl start mysql
+ECHO MySql server started
+read
