--- scripts/!quickStartLin/1-startOrientDb.sh
@@ -1,1 +1,0 @@
-sh /opt/orientdb/bin/server.sh
