--- scripts/!quickStartLin/2-initDb.sh
@@ -1,1 +1,0 @@
-/opt/orientdb/bin/console.sh /www/codeforthefences/scripts/osql/resetDb.sh.osql
