+++ scripts/!quickStartLin/2-resetDb.sh
@@ -2,8 +2,8 @@
 ECHO Running scripts
 mysql -u root -ppassword < C:/node/codeforthefences/scripts/mysql/resetDb.sql
 ECHO Database dropped/created
-mysql -u root -ppassword < C:/node/codeforthefences/scripts/mysql/graphDesign.sql
+mysql -u root -ppassword < C:/node/codeforthefences/scripts/mysql/design.sql
 ECHO Database design added
-mysql -u root -ppassword < C:/node/codeforthefences/scripts/mysql/graphSampleData_Base.sql
+mysql -u root -ppassword < C:/node/codeforthefences/scripts/mysql/sampleData.sql
 ECHO Database sample data added
-pause
+read
