+++ scripts/!quickStartWin/1-launchDbService.bat
@@ -1,4 +1,3 @@
 @echo off
-REM %CFTF_HOME%/scripts/!quickStartWin/osql/1-startOrientDb.bat
-
-runas /user:Administrator %CFTF_HOME%/scripts/!quickStartWin/mysql/1-startMySql.bat
+runas /user:heDev %CFTF_HOME%/scripts/!quickStartWin/mysql/1-startMySql.bat
+pause
