--- scripts/!quickStartWin/2-initDb.bat
@@ -1,2 +1,0 @@
-REM %CFTF_HOME%/scripts/!quickStartWin/osql/2-initOrientDb.bat
-%CFTF_HOME%/scripts/!quickStartWin/mysql/2-initMySqlDb.bat
