+++ scripts/!quickStartWin/mysql/1-startMySql.bat
@@ -1,6 +1,4 @@
 @ECHO OFF
-REM "C:\Program Files\MySQL\MySQL Server 8.0\bin\mysqladmin" -u root shutdown
-
 ECHO Starting MySql server
 NET START MySQL80
 ECHO MySql server started
