--- scripts/!quickStartWin/osql/1-startOrientDb.bat
@@ -1,1 +1,0 @@
-CALL %ORIENTDB_HOME%\bin\server.bat
