--- scripts/!quickStartWin/osql/2-initOrientDb.bat
@@ -1,2 +1,0 @@
-CALL %ORIENTDB_HOME%\bin\console.bat %CFTF_HOME%\scripts\osql\resetDb.osql
-pause
