+++ scripts/mysql/design.sql
@@ -2,6 +2,7 @@ USE codeforthefences;
 
 CREATE TABLE user (
   `id` int AUTO_INCREMENT primary key NOT NULL,
+  `createdDate` datetime NOT NULL,
   `username` nvarchar(64) UNIQUE NOT NULL,
   `password` nvarchar(128) NOT NULL,
   `sessionId` nvarchar(128) NULL,
@@ -10,8 +11,9 @@ CREATE TABLE user (
 
 CREATE TABLE blogPost (
   `id` int AUTO_INCREMENT primary key NOT NULL,
+  `authorId` int NOT NULL,
   `title` nvarchar(64) NOT NULL,
-  `date` datetime NOT NULL,
+  `createdDate` datetime NOT NULL,
   `content` mediumtext NOT NULL
 );
 
@@ -27,32 +29,31 @@ CREATE TABLE toolType (
   `name` nvarchar(64) NOT NULL,
   `description` nvarchar(256) NOT NULL
 );
-
-CREATE TABLE UoM (
-  `id` int AUTO_INCREMENT primary key NOT NULL,
-  `name` nvarchar(64) NOT NULL,
-  `abbreviation` nvarchar(32) NOT NULL
-);
-
 CREATE TABLE tool (
   `id` int AUTO_INCREMENT primary key NOT NULL,
   `toolTypeId` int NOT NULL,
   `UoMId` int NULL
 );
-
-CREATE TABLE recipeTool (
+CREATE TABLE recipe_tool (
   `id` int AUTO_INCREMENT primary key NOT NULL,
   `recipeId` int NOT NULL,
   `toolId` int NOT NULL,
   `toolIndex` int NOT NULL
 );
 
-CREATE TABLE foodType (
+CREATE TABLE ingredient (
   `id` int AUTO_INCREMENT primary key NOT NULL,
-  `name` nvarchar(32) NOT NULL UNIQUE,
-  `plural` nvarchar(32) NULL,
-  `abbreviation` nvarchar(16) NULL,
-  `plAbbrev` nvarchar(16) NULL
+  `UoMId` int NULL,
+  `foodVariantId` int NULL,
+  `substanceId` int NOT NULL,
+  `prepStyleId` int NULL
+);
+
+CREATE TABLE UoM (
+  `id` int AUTO_INCREMENT primary key NOT NULL,
+  `name` nvarchar(64) NOT NULL,
+  `plural` nvarchar(64) NULL,
+  `abbreviation` nvarchar(32) NULL
 );
 
 CREATE TABLE foodVariant (
@@ -62,22 +63,12 @@ CREATE TABLE foodVariant (
   `description` nvarchar(64) NULL
 );
 
-CREATE TABLE food (
+CREATE TABLE substance (
   `id` int AUTO_INCREMENT primary key NOT NULL,
-  `foodTypeId` int NOT NULL,
-  `foodVariantId` int NULL,
-  CONSTRAINT uc_food UNIQUE (foodTypeId, foodVariantId)
-);
-
-CREATE TABLE foodCategory (
-  `id` int AUTO_INCREMENT primary key NOT NULL,
-  `name` nvarchar(64) NOT NULL
-);
-
-CREATE TABLE categorizedFood (
-  `id` int AUTO_INCREMENT primary key NOT NULL,
-  `foodId` int NOT NULL,
-  `foodCategoryId` int NOT NULL
+  `name` nvarchar(32) NOT NULL UNIQUE,
+  `plural` nvarchar(32) NULL,
+  `abbreviation` nvarchar(16) NULL,
+  `plAbbrev` nvarchar(16) NULL
 );
 
 CREATE TABLE prepStyle (
@@ -86,29 +77,12 @@ CREATE TABLE prepStyle (
   `description` nvarchar(64) NOT NULL
 );
 
-CREATE TABLE preppedFood (
-  `id` int AUTO_INCREMENT primary key NOT NULL,
-  `foodId` int NOT NULL,
-  `prepStyleId` int NULL
-);
-
-CREATE TABLE measureOfFood (
-  `id` int AUTO_INCREMENT primary key NOT NULL,
-  `preppedFoodId` int NULL,
-  `UoMId` int NOT NULL
-);
-
-CREATE TABLE ingredient (
-  `id` int AUTO_INCREMENT primary key NOT NULL,
-  `measureOfFoodId` int NOT NULL,
-  `quantity` float NOT NULL
-);
-
-CREATE TABLE recipeIngredient (
+CREATE TABLE recipe_ingredient (
   `id` int AUTO_INCREMENT primary key NOT NULL,
   `recipeId` int NOT NULL,
   `ingredientId` int NOT NULL,
-  `ingredientIndex` int NOT NULL
+  `ingredientIndex` int NOT NULL,
+  `quantity` float NOT NULL
 );
 
 CREATE TABLE step (
@@ -116,23 +90,22 @@ CREATE TABLE step (
   `text` nvarchar(4000) NOT NULL
 );
 
-CREATE TABLE recipeStep (
+CREATE TABLE recipe_step (
   `id` int AUTO_INCREMENT primary key NOT NULL,
   `recipeId` int NOT NULL,
   `stepId` int NOT NULL,
   `stepIndex` int NOT NULL
 );
 
-CREATE TABLE recipeStepTool (
+CREATE TABLE stepMapType (
   `id` int AUTO_INCREMENT primary key NOT NULL,
-  `recipeStepId` int NOT NULL,
-  `barsIndex` int NOT NULL,
-  `recipeToolIndex` int NOT NULL
+  `mapType` nvarchar(32) NOT NULL
 );
 
-CREATE TABLE recipeStepIngredient (
+CREATE TABLE stepMap (
   `id` int AUTO_INCREMENT primary key NOT NULL,
   `recipeStepId` int NOT NULL,
+  `stepMapTypeId` int NOT NULL,
   `barsIndex` int NOT NULL,
-  `recipeIngredientIndex` int NOT NULL
+  `recipeIndex` int NOT NULL
 );
