+++ scripts/mysql/sampleData.sql
@@ -52,6 +52,16 @@ SET @creamId = LAST_INSERT_ID();
 
 INSERT INTO substance (name) VALUES ('parsley');
 INSERT INTO substance (name) VALUES ('sausage');
+INSERT INTO foodVariant (name) VALUES ('chicken');
+SET @chickenId = LAST_INSERT_ID();
+INSERT INTO foodVariant (name) VALUES ('peas');
+SET @peasId = LAST_INSERT_ID();
+INSERT INTO foodVariant (name) VALUES ('blackened');
+SET @blackenedId = LAST_INSERT_ID();
+INSERT INTO foodVariant (name) VALUES ('canned');
+SET @cannedId = LAST_INSERT_ID();
+INSERT INTO foodVariant (name) VALUES ('mushy');
+SET @mushyId = LAST_INSERT_ID();
 
 INSERT INTO foodVariant (name) VALUES ('mini');
 SET @miniId = LAST_INSERT_ID();
