--- scripts/osql/loadSampleData_base.bat
@@ -1,2 +1,0 @@
-CALL %ORIENTDB_HOME%/console.bat %CODEFORTHEFENCES_HOME%scripts/osql/loadSampleData_base.osql
-pause
