--- scripts/osql/resetDb.osql
@@ -1,11 +1,0 @@
-SET echo TRUE;
-SET promptDateFormat "yyy-MM-dd hh:mm:ss.sss";
-CONNECT remote:localhost root testpass123;
-LIST DATABASES;
-SET ignoreErrors TRUE;
-DROP DATABASE remote:localhost/codeforthefences root testpass123;
-SET ignoreErrors FALSE;
-CREATE DATABASE remote:localhost/codeforthefences root testpass123 PLOCAL;
-
-LOAD SCRIPT C:/node/codeforthefences/scripts/osql/schema.osql;
-LOAD SCRIPT C:/node/codeforthefences/scripts/osql/sampleData_base.osql;
